rename ws client main.py to client.py
