+++ producer/main.py
@@ -1,9 +1,9 @@
 """Main executable of the LOVE-producer."""
 import asyncio
-from telemetries_events.main import main as telemetries_events
-from scriptqueue.main import main as scriptqueue
-from heartbeats.main import main as heartbeats
-from command_receiver.main import main as command_receiver
+from telemetries_events.client import main as telemetries_events
+from scriptqueue.client import main as scriptqueue
+from heartbeats.client import main as heartbeats
+from command_receiver.client import main as command_receiver
 
 if __name__ == '__main__':
     loop = asyncio.get_event_loop()
